+++ KmpCrypto/src/jvmTest/kotlin/com/oldguy/crypto/FileTests.kt
@@ -1,7 +1,6 @@
 package com.oldguy.crypto
 
 import com.oldguy.common.io.*
-import com.oldguy.common.toHex
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
@@ -9,11 +8,24 @@ import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
+/**
+ * Uses a zip file (zip64 spec) as a source payload for all encryption tests, as makes content of
+ * large file easy to verify.
+ *
+ * Currently only test is AES/GCM, need to add tests of same process using additional algorithms.
+ */
 @ExperimentalCoroutinesApi
 class FileTests {
     private val testZipName = "ZerosZip64.zip"
     val password = "xyzAbc!@#"
+    val zipSize = 5611756UL
+    val encryptedSize = (5611756 + 13 + 16).toULong()
 
+    /**
+     * Encrypt a text zip64 file, using AES and GCM and a random IV. Encrypt same file with bouncy
+     * castle, compare encrypted files to ensure size and content match.  Decrypt file. Verify Zip
+     * file decrypted content using Zip support.
+     */
     @Test
     fun fileAesGcm() {
         val testDirectory = File("TestFiles")
@@ -34,37 +46,53 @@ class FileTests {
             }
             val ivContent = cipher.keyConfiguration.iv
             val source = RawFile(zipFile)
-            val dest = RawFile(zipEncrypted, FileMode.Write).apply {
-                UByteBuffer(cipher.keyConfiguration.ivSize + 1).apply {
-                    byte = cipher.engine.ivSize.toUByte()
-                    putBytes(cipher.keyConfiguration.iv)
-                    flip()
-                    write(this)
-                }
-            }
+            val dest = RawFile(zipEncrypted, FileMode.Write)
+            cipher.writeInitializationVector(dest)
             cipher.process(true, source, dest)
             assertTrue(zipEncrypted.exists)
+            assertEquals(zipSize, zipFile.size)
+            assertEquals(encryptedSize, zipEncrypted.size)
 
-            val b = bouncyEncrypt(workDirectory, zipFile, ivContent)
-            assertTrue(CryptoTestHelp.compare(zipEncrypted, b))
+            // match to bouncycastle copy
+            val bouncy = bouncyEncrypt(workDirectory, zipFile, ivContent)
+            assertEquals(encryptedSize, bouncy.size)
+            assertTrue(CryptoTestHelp.compare(zipEncrypted, bouncy))
 
+            // verify IV content
             val check = RawFile(zipEncrypted)
-            val buf = ByteBuffer(1)
-            assertEquals(1u, check.read(buf))
-            buf.flip()
-            val length = buf.byte.toUInt()
-            assertEquals(12u, length)
-            val ivBuf = UByteBuffer(length.toInt())
-            assertEquals(length, check.read(ivBuf))
-            assertContentEquals(ivContent, ivBuf.flip().getBytes())
+            val ivRead = cipher.readInitializationVector(check)
+            assertContentEquals(ivContent, ivRead)
 
+            // decrypt
             val decryptFile = File(workDirectory, "$testZipName.decrypt.zip")
             decryptFile.delete()
             val zipDecrypted = RawFile(decryptFile, FileMode.Write)
             cipher.process(false, check, zipDecrypted)
+
+            // Verify zip content, which is Zip64 one file expands to 5GB of zeroes. Extract not done,
+            // just verify directory content (which is near end of zip file anyway).
+            ZipFile(decryptFile).apply {
+                open()
+                val entryName = "0000"
+                assertEquals(1, map.size)
+                assertEquals(entryName, map.keys.first())
+                val entry = map[entryName]
+                    ?: throw IllegalStateException("Entry $entryName not found")
+                assertEquals(entryName, entry.name)
+                assertEquals(5611526UL, entry.directories.compressedSize)
+                assertEquals(5368709120UL, entry.directories.uncompressedSize)
+            }
+            decryptFile.delete()
+            bouncy.delete()
+            zipEncrypted.delete()
         }
     }
 
+    /**
+     * Use same IV and algorithms to encrypt same target file using Bouncy Castle.
+     * Note as part of this test a BufferReader was used solely as a larger test of BufferReader,
+     * not necessary for bouncycastle use.
+     */
     suspend fun bouncyEncrypt(
         workDirectory: File,
         source: File,
@@ -97,10 +125,9 @@ class FileTests {
         var bytesIn = 0UL
         var blocksRead = 0
         val reader = BufferReader {
-            payload.clear()
-            bytesIn += zip.read(payload)
+            bytesIn += zip.read(payload, true)
             blocksRead++
-            payload.flip()
+            payload
         }
         val blockOut = ByteArray(eng.blockSize)
         var totalLength = 0